+++ packages/local-prover/CHANGELOG.md
@@ -10,7 +10,7 @@
 - add `encSenderCanonAddrC1X`, `encSenderCanonAddrC2X` to public signals
 - add `encRandomness` to inputs
 - clean up module hierarchy, remove unnecessary exports, and make exports explicit
-- use `poseidonBN` from `@nocturne-xyz/circuit-utils`
+- use `poseidonBN` from `@nocturne-xyz/crypto-utils`
 - Rename 
 	- `LocalJoinSplitProver` -> `WasmJoinSplitProver`
 	- `LocalSubtreeUpdateProver` -> `WasmSubtreeUpdateProver`